chore: テーマカラー周りの修正
