+++ src/components/elements/SideBar/SideBarItem/SideBarItem.tsx
@@ -3,7 +3,7 @@ import { ListItemButton, ListItemIcon, Typography } from '@mui/material'
 import { useAtom } from 'jotai'
 import { useRouter } from 'next/router'
 import { memo, useCallback } from 'react'
-import { colorConfigs } from 'src/config/color'
+import { colorConfigs } from 'src/config/theme'
 import { SideBarNavigationsType } from 'src/domain/sideBar/sideBar'
 import { pagesPath } from 'src/lib/$path'
 import { activeSideBarItemAtom } from 'src/stores/atom'
