--- src/config/color.ts
@@ -1,12 +1,0 @@
-export const colorConfigs = {
-  sidebar: {
-    activeBg: '#ff79c6',
-  },
-  text: {
-    primary: '#f8f8f2',
-    secondary: '#bd93f9',
-    tertiary: '#a1a2a7',
-    quaternary: '#48bd67',
-    quinary: '#ff79c6',
-  },
-}
