+++ src/config/theme.ts
@@ -14,6 +14,19 @@ export const theme = createTheme({
   },
 })
 
+export const colorConfigs = {
+  sidebar: {
+    activeBg: '#ff79c6',
+  },
+  text: {
+    primary: '#f8f8f2',
+    secondary: '#bd93f9',
+    tertiary: '#a1a2a7',
+    quaternary: '#48bd67',
+    quinary: '#ff79c6',
+  },
+}
+
 export const getColorTheme = (mode: ThemeColorType) => {
   return createTheme({
     typography: {
